Moved SaleForce cache stuff to its own package.
