+++ orcid-core/src/main/java/org/orcid/core/salesforce/cache/MemberDetailsCacheEntryFactory.java
@@ -14,11 +14,12 @@
  *
  * =============================================================================
  */
-package org.orcid.core.manager.impl;
+package org.orcid.core.salesforce.cache;
 
 import javax.annotation.Resource;
 
 import org.orcid.core.salesforce.dao.SalesForceDao;
+import org.orcid.core.salesforce.model.MemberDetails;
 
 import net.sf.ehcache.constructs.blocking.CacheEntryFactory;
 
@@ -27,14 +28,16 @@ import net.sf.ehcache.constructs.blocking.CacheEntryFactory;
  * @author Will Simpson
  *
  */
-public class SalesForceMembersListCacheEntryFactory implements CacheEntryFactory {
+public class MemberDetailsCacheEntryFactory implements CacheEntryFactory {
 
     @Resource
     private SalesForceDao salesForceDao;
 
     @Override
     public Object createEntry(Object key) throws Exception {
-        return salesForceDao.retrieveFreshMembers();
+        MemberDetailsCacheKey detailsCacheKey = (MemberDetailsCacheKey) key;
+        MemberDetails details = salesForceDao.retrieveFreshDetails(detailsCacheKey.getMemberId(), detailsCacheKey.getConsotiumLeadId());
+        return details;
     }
 
 }
