+++ orcid-core/src/main/java/org/orcid/core/salesforce/cache/MembersListCacheEntryFactory.java
@@ -14,12 +14,11 @@
  *
  * =============================================================================
  */
-package org.orcid.core.manager.impl;
+package org.orcid.core.salesforce.cache;
 
 import javax.annotation.Resource;
 
 import org.orcid.core.salesforce.dao.SalesForceDao;
-import org.orcid.core.salesforce.model.MemberDetails;
 
 import net.sf.ehcache.constructs.blocking.CacheEntryFactory;
 
@@ -28,16 +27,14 @@ import net.sf.ehcache.constructs.blocking.CacheEntryFactory;
  * @author Will Simpson
  *
  */
-public class SalesForceMemberDetailsCacheEntryFactory implements CacheEntryFactory {
+public class MembersListCacheEntryFactory implements CacheEntryFactory {
 
     @Resource
     private SalesForceDao salesForceDao;
 
     @Override
     public Object createEntry(Object key) throws Exception {
-        SalesForceMemberDetailsCacheKey detailsCacheKey = (SalesForceMemberDetailsCacheKey) key;
-        MemberDetails details = salesForceDao.retrieveFreshDetails(detailsCacheKey.getMemberId(), detailsCacheKey.getConsotiumLeadId());
-        return details;
+        return salesForceDao.retrieveFreshMembers();
     }
 
 }
