+++ orcid-core/src/main/resources/orcid-core-context.xml
@@ -607,7 +607,7 @@
         <property name="copyOnRead" value="false" />
         <property name="copyOnWrite" value="false" />
         <property name="cacheEntryFactory">
-            <bean class="org.orcid.core.manager.impl.SalesForceMembersListCacheEntryFactory" />
+            <bean class="org.orcid.core.salesforce.cache.MembersListCacheEntryFactory" />
         </property>
     </bean>
     
@@ -621,7 +621,7 @@
         <property name="copyOnRead" value="false" />
         <property name="copyOnWrite" value="false" />
         <property name="cacheEntryFactory">
-            <bean class="org.orcid.core.manager.impl.SalesForceMemberDetailsCacheEntryFactory" />
+            <bean class="org.orcid.core.salesforce.cache.MemberDetailsCacheEntryFactory" />
         </property>
     </bean>
